+++ src/main/java/com/fastcampus/shop/controller/QnaController.java
@@ -1,10 +1,10 @@
 package com.fastcampus.shop.controller;
 
 
-import com.fastcampus.shop.dto.PageHandler;
+import com.fastcampus.shop.domain.QPageHandler;
+import com.fastcampus.shop.domain.QSearchCondition;
 import com.fastcampus.shop.dto.QnaCommentDto;
 import com.fastcampus.shop.dto.QnaDto;
-import com.fastcampus.shop.dto.SearchCondition;
 import com.fastcampus.shop.service.AdminService;
 import com.fastcampus.shop.service.QnaCommentService;
 import com.fastcampus.shop.service.QnaService;
@@ -120,7 +120,7 @@ public class QnaController {
     @PostMapping("/remove")
     public String remove(@RequestParam("qnaId") Integer qnaId,
                          /*@RequestParam("memberId") Integer memberId,*/
-                         SearchCondition sc, Model m, HttpSession session, RedirectAttributes rattr) {
+                         QSearchCondition sc, Model m, HttpSession session, RedirectAttributes rattr) {
         // 테스트용 memberId 하드코딩
         //Integer memberId = 1001; // ← 실제 존재하는 FK 값
 
@@ -170,7 +170,7 @@ public class QnaController {
     }
 
     @GetMapping("/read")
-    public String read(Integer qnaId, SearchCondition sc, Model m, HttpSession session, RedirectAttributes rattr) {
+    public String read(Integer qnaId, QSearchCondition sc, Model m, HttpSession session, RedirectAttributes rattr) {
 
         //session.setAttribute("memberId", 1001);
         Integer memberId = (Integer) session.getAttribute("memberId");
@@ -195,7 +195,7 @@ public class QnaController {
     }
 
     @GetMapping("/list")
-    public String list(SearchCondition sc, Model m, HttpServletRequest request) {
+    public String list(QSearchCondition sc, Model m, HttpServletRequest request) {
         if (!loginCheck(request))
             return "redirect:/login/login?toURL=" + request.getRequestURL();  // 로그인을 안했으면 로그인 화면으로 이동
         System.out.println("### sc.getCurrentPage() = " + sc.getCurrentPage());
@@ -205,7 +205,7 @@ public class QnaController {
             int totalCnt = qnaService.getSearchResultCnt(sc);
             m.addAttribute("totalCnt", totalCnt);
 
-            PageHandler pageHandler = new PageHandler(totalCnt, sc);
+            QPageHandler pageHandler = new QPageHandler(totalCnt, sc);
 
             List<QnaDto> list = qnaService.getSearchResultPage(sc);
             //List<QnaDto> list = qnaService.getSearchResultPageWithMemberName(sc);
@@ -231,5 +231,3 @@ public class QnaController {
         return true;  // 로그인 했다고 가정
     }
 }
-
-
