+++ src/main/java/com/fastcampus/shop/dto/Member.java
@@ -1,0 +1,25 @@
+package com.fastcampus.shop.dto;
+
+public class Member {
+    private Long memberId;
+    private String memberLoginId;
+    private String name;
+
+    // Getters and Setters
+    public Long getMemberId() {
+        return memberId;
+    }
+
+    public void setMemberId(Long memberId) {
+        this.memberId = memberId;
+    }
+
+    public String getMemberLoginId() {
+        return memberLoginId;
+    }
+
+    public void setMemberLoginId(String memberLoginId) {
+        this.memberLoginId = memberLoginId;
+    }
+
+}
