+++ src/main/java/com/fastcampus/shop/dto/QnaDto.java
@@ -16,11 +16,10 @@ public class QnaDto {
     private Date updateAt;
     private int viewCnt;
     private int commentCnt;
-    /*private String memberName;
+    private String memberName;
 
     public String getMemberName() { return memberName; }
     public void setMemberName(String memberName) { this.memberName = memberName; }
-*/
 
     public Integer getQnaId() {
         return qnaId;
@@ -148,4 +147,3 @@ public class QnaDto {
         return Objects.hash(qnaId, memberId, title, content, qnaCategory, isSecret, productId, password);
     }
 }
-
