+++ src/main/java/com/fastcampus/shop/service/QMemberService.java
@@ -1,0 +1,5 @@
+package com.fastcampus.shop.service;
+
+public interface QMemberService {
+    Long getMemberIdByLoginId(String memberLoginId);
+}
