+++ src/main/resources/mappers/QmemberMapper.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.fastcampus.shop.dao.MemberMapper">
+
+    <select id="findMemberIdByLoginId" parameterType="String" resultType="long">
+        SELECT member_id
+        FROM member
+        WHERE member_login_id = #{memberLoginId}
+    </select>
+
+</mapper>
