+++ src/main/resources/mappers/QnaMapper.xml
@@ -6,21 +6,24 @@
 
     <!-- 전체 조회 -->
     <select id="findAll" resultType="com.fastcampus.shop.dto.QnaDto">
-        SELECT qna_id, member_id, title, content, create_at, view_cnt, comment_cnt, update_at
-        FROM qna
-        ORDER BY create_at DESC, qna_id DESC <!-- 최신순으로 정렬 -->
+        SELECT q.qna_id, q.member_id, q.title, q.content, q.create_at, q.view_cnt, q.comment_cnt, q.update_at, m.member_name AS member_name
+        FROM qna q
+        JOIN member m ON q.member_id = m.member_id
+        ORDER BY q.create_at DESC, q.qna_id DESC <!-- 최신순으로 정렬 -->
     </select>
     <!-- 상세 조회 -->
     <select id="findById" parameterType="int" resultType="com.fastcampus.shop.dto.QnaDto">
-        SELECT qna_id, member_id, title, content, create_at, view_cnt, comment_cnt, update_at
-        FROM qna
-        WHERE qna_id = #{qnaId}
+        SELECT q.qna_id, q.member_id, q.title, q.content, q.create_at, q.view_cnt, q.comment_cnt, q.update_at, m.member_name AS member_name
+        FROM qna q
+        JOIN member m ON q.member_id = m.member_id
+        WHERE q.qna_id = #{qnaId}
     </select>
     <!-- 사용자별 상세 조회 -->
     <select id="findByMemberId" parameterType="int" resultType="com.fastcampus.shop.dto.QnaDto">
-        SELECT qna_id, member_id, title, content, create_at, view_cnt, comment_cnt, update_at
-        FROM qna
-        WHERE member_id = #{memberId}
+        SELECT q.qna_id, q.member_id, q.title, q.content, q.create_at, q.view_cnt, q.comment_cnt, q.update_at, m.member_name AS member_name
+        FROM qna q
+        JOIN member m ON q.member_id = m.member_id
+        WHERE q.member_id = #{memberId}
     </select>
     <!-- 등록 -->
     <insert id="insert" parameterType="com.fastcampus.shop.dto.QnaDto" useGeneratedKeys="true" keyProperty="qnaId">
@@ -61,9 +64,10 @@
 
     <!-- 페이징된 게시글 목록 조회 -->
     <select id="selectPage" parameterType="map" resultType="com.fastcampus.shop.dto.QnaDto">
-        SELECT qna_id, member_id, title, content, create_at
-        FROM qna
-        ORDER BY create_at DESC, qna_id DESC
+        SELECT q.qna_id, q.member_id, q.title, q.content, q.create_at, q.view_cnt, q.comment_cnt, m.member_name AS member_name
+        FROM qna q
+        JOIN member m ON q.member_id = m.member_id
+        ORDER BY q.create_at DESC, q.qna_id DESC
         LIMIT #{offset}, #{pageSize}
     </select>
 
@@ -73,7 +77,7 @@
                 AND title LIKE concat('%', #{keyword}, '%')
             </when>
             <when test='option=="W"'>
-                AND member_id LIKE concat('%', #{keyword}, '%')
+                AND m.member_name LIKE concat('%', #{keyword}, '%')
             </when>
             <otherwise>
                 AND (title LIKE concat('%', #{keyword}, '%')
@@ -83,17 +87,19 @@
     </sql>
 
     <select id="searchSelectPage" parameterType="SearchCondition" resultType="com.fastcampus.shop.dto.QnaDto">
-        SELECT qna_id, member_id, title, content, create_at, view_cnt, comment_cnt
-        FROM qna
+        SELECT q.qna_id, q.member_id, q.title, q.content, q.create_at, q.view_cnt, q.comment_cnt, m.member_name AS member_name
+        FROM qna q
+        JOIN member m ON q.member_id = m.member_id
         WHERE true
         <include refid="searchCondition"/>
-        ORDER BY create_at DESC, qna_id DESC
+        ORDER BY q.create_at DESC, q.qna_id DESC
         LIMIT #{offset}, #{pageSize}
     </select>
 
     <select id="searchResultCnt" parameterType="SearchCondition" resultType="int">
         SELECT count(*)
-        FROM qna
+        FROM qna q
+        JOIN member m ON q.member_id = m.member_id
         WHERE true
         <include refid="searchCondition"/>
     </select>
@@ -122,11 +128,11 @@
     </delete>
 
     <!-- 게시글에 작성자 이름까지 같이 조회 -->
-    <!--<select id="findAllWithMemberName" resultType="com.fastcampus.shop.dto.QnaDto">
-        SELECT q.qna_id, q.title, q.content, q.create_at, q.member_id, m.name AS member_name
+    <select id="findAllWithMemberName" resultType="com.fastcampus.shop.dto.QnaDto">
+        SELECT q.qna_id, q.title, q.content, q.create_at, q.view_cnt, q.comment_cnt, q.member_id, m.member_name AS member_name
         FROM qna q
                  JOIN member m ON q.member_id = m.member_id
-        ORDER BY q.create_at DESC
-    </select>-->
+        ORDER BY q.create_at DESC, q.qna_id DESC
+    </select>
 
 </mapper>
