+++ src/main/webapp/resources/css/qna.css
@@ -1,0 +1,451 @@
+/* QnA CSS Styles */
+body {
+    font-family: 'Noto Sans KR', sans-serif;
+    background-color: #f9fafb;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+/* Common container styles */
+.container, .qna-container {
+    max-width: 1200px;
+    margin: 0 auto;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+/* Header styles */
+.writing-header {
+    text-align: center;
+    font-size: 28px;
+    font-weight: bold;
+    margin-top: 20px;
+    margin-bottom: 30px;
+    color: #333;
+}
+
+/* QnA Header styles */
+.qna-header {
+    text-align: center;
+    width: 100%;
+    max-width: 1000px;
+    margin-bottom: 40px;
+    padding: 20px 0;
+}
+
+.qna-title {
+    font-size: 42px;
+    font-weight: bold;
+    margin-bottom: 30px;
+    color: #333;
+}
+
+.qna-info {
+    font-size: 14px;
+    line-height: 1.6;
+    color: #666;
+    margin: 0 auto;
+    max-width: 600px;
+}
+
+/* Form styles */
+.frm {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    max-width: 1000px;
+    margin: 0 auto;
+    background-color: white;
+    padding: 30px;
+    border-radius: 8px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.frm input[type="text"], .frm textarea {
+    width: 100%;
+    padding: 12px;
+    margin-bottom: 20px;
+    box-sizing: border-box;
+    font-size: 16px;
+    border: 1px solid #d1d5db;
+    border-radius: 4px;
+    transition: border-color 0.3s, box-shadow 0.3s;
+}
+
+.frm input[type="text"]:focus, .frm textarea:focus {
+    outline: none;
+    border-color: #a855f7;
+    box-shadow: 0 0 0 3px rgba(168, 85, 247, 0.3);
+}
+
+.frm textarea {
+    height: 300px;
+    resize: vertical;
+}
+
+/* Button styles */
+.button-row {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 15px;
+    margin-top: 20px;
+    width: 100%;
+}
+
+.btn {
+    padding: 8px 16px;
+    font-size: 16px;
+    font-weight: 500;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+}
+
+.btn-write {
+    background-color: #a855f7;
+    color: white;
+}
+
+.btn-write:hover {
+    background-color: #9333ea;
+}
+
+.btn-modify {
+    background-color: #3b82f6;
+    color: white;
+}
+
+.btn-modify:hover {
+    background-color: #2563eb;
+}
+
+.btn-remove {
+    background-color: #ef4444;
+    color: white;
+}
+
+.btn-remove:hover {
+    background-color: #dc2626;
+}
+
+.btn-list {
+    background-color: #6b7280;
+    color: white;
+}
+
+.btn-list:hover {
+    background-color: #4b5563;
+}
+
+/* QnA List Table Styles */
+table {
+    width: 100%;
+    max-width: 1000px;
+    border-collapse: collapse;
+    margin: 20px 0;
+    background-color: white;
+    border-radius: 8px;
+    overflow: hidden;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+th, td {
+    padding: 15px;
+    text-align: left;
+    border-bottom: 1px solid #e5e7eb;
+}
+
+th {
+    background-color: #f3f4f6;
+    font-weight: bold;
+    color: #374151;
+}
+
+tr:hover {
+    background-color: #f9fafb;
+}
+
+.no {
+    width: 10%;
+    text-align: center;
+}
+
+.title {
+    width: 50%;
+}
+
+.writer {
+    width: 15%;
+    text-align: center;
+}
+
+.createAt, .regdate {
+    width: 15%;
+    text-align: center;
+}
+
+.viewCnt, .viewcnt {
+    width: 10%;
+    text-align: center;
+}
+
+table a {
+    color: #374151;
+    text-decoration: none;
+    font-weight: 500;
+}
+
+table a:hover {
+    color: #a855f7;
+    text-decoration: underline;
+}
+
+/* Search container styles */
+.search-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    max-width: 1000px;
+    margin-bottom: 20px;
+}
+
+.search-form {
+    display: flex;
+    gap: 10px;
+}
+
+.search-option {
+    padding: 8px 12px;
+    border: 1px solid #d1d5db;
+    border-radius: 4px;
+    background-color: white;
+}
+
+.search-input {
+    padding: 8px 12px;
+    border: 1px solid #d1d5db;
+    border-radius: 4px;
+    width: 300px;
+}
+
+.search-button {
+    padding: 8px 16px;
+    background-color: #a855f7;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+.search-button:hover {
+    background-color: #9333ea;
+}
+
+/* Pagination styles */
+.paging-container {
+    display: flex;
+    justify-content: center;
+    margin: 30px 0;
+}
+
+.paging {
+    display: flex;
+    gap: 10px;
+}
+
+.page {
+    display: inline-block;
+    padding: 8px 12px;
+    border: 1px solid #d1d5db;
+    border-radius: 4px;
+    color: #374151;
+    text-decoration: none;
+    transition: background-color 0.3s;
+}
+
+.page:hover {
+    background-color: #f3f4f6;
+}
+
+.paging-active {
+    background-color: #a855f7;
+    color: white;
+    border-color: #a855f7;
+}
+
+.paging-active:hover {
+    background-color: #9333ea;
+}
+
+/* Comment section styles */
+.comment-section {
+    width: 100%;
+    max-width: 1000px;
+    margin: 30px auto;
+    background-color: white;
+    padding: 30px;
+    border-radius: 8px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+#commentList {
+    width: 100%;
+    margin-top: 30px;
+}
+
+#commentList ul {
+    list-style-type: none;
+    padding: 0;
+}
+
+#commentList li {
+    padding: 15px;
+    border-bottom: 1px solid #e5e7eb;
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    gap: 10px;
+}
+
+#commentList .commenter {
+    font-weight: bold;
+    color: #374151;
+}
+
+#commentList .content {
+    flex-grow: 1;
+}
+
+/* Style for empty comment container to maintain consistent width */
+.empty-comment-container {
+    width: 100%;
+    min-height: 50px;
+    border-bottom: 1px solid #e5e7eb;
+}
+
+#commentList button {
+    padding: 6px 12px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    margin-left: 5px;
+}
+
+#commentList .delBtn {
+    background-color: #ef4444;
+    color: white;
+}
+
+#commentList .modBtn {
+    background-color: #3b82f6;
+    color: white;
+}
+
+#commentList .replyBtn {
+    background-color: #10b981;
+    color: white;
+}
+
+/* Comment input area */
+#comment {
+    width: 70%;
+    padding: 10px;
+    border: 1px solid #d1d5db;
+    border-radius: 4px;
+    margin-right: 10px;
+}
+
+#sendBtn, #modBtn {
+    padding: 8px 16px;
+    background-color: #a855f7;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+#sendBtn:hover, #modBtn:hover {
+    background-color: #9333ea;
+}
+
+/* Reply form */
+#replyForm {
+    margin-top: 10px;
+    margin-left: 30px;
+    padding: 10px;
+    background-color: #f3f4f6;
+    border-radius: 4px;
+}
+
+#replyForm input[type="text"] {
+    width: 70%;
+    padding: 8px;
+    border: 1px solid #d1d5db;
+    border-radius: 4px;
+    margin-right: 10px;
+}
+
+#wrtRepBtn {
+    padding: 8px 16px;
+    background-color: #10b981;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+#wrtRepBtn:hover {
+    background-color: #059669;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+    .frm {
+        width: 90%;
+        padding: 20px;
+    }
+
+    .search-container {
+        flex-direction: column;
+        gap: 15px;
+    }
+
+    .search-form {
+        width: 100%;
+    }
+
+    .search-input {
+        width: 100%;
+    }
+
+    table {
+        font-size: 14px;
+    }
+
+    th, td {
+        padding: 10px;
+    }
+
+    .no, .viewCnt, .viewcnt {
+        display: none;
+    }
+
+    .title {
+        width: 60%;
+    }
+
+    .writer {
+        width: 20%;
+    }
+
+    .createAt, .regdate {
+        width: 20%;
+    }
+}
